chore: service implementation and bank edit

+++ ExpenseTracker/Models/Bank.cs
@@ -10,7 +10,7 @@ namespace ExpenseTracker.Models
         [Key] [Column("id")] public int Id { get; set; }
         [Column("bankname")] public string BankName { get; set; }
         [Column("accountnumber")] public string AccountNumber { get; set; }
-        [Column("bankcontactnumber")] public int BankContactNumber { get; set; }
+        [Column("bankcontactnumber")] public long BankContactNumber { get; set; }
         [Column("ledgerid")] public int LedgerId { get; set; }
         [Column("remainingbalance")] public decimal RemainingBalance { get; set; }
         [Column("bankaddress")] public string BankAddress { get; set; }
